basic migration of logic from tacker component
