+++ src/locationTracker/LocationTracker.js
@@ -8,12 +8,12 @@ class LocationTracker extends React.Component {
         return (
             <div className="location-tracker" style={this.props.style}>
                 <ul style={{padding: "2%"}}>
-                    {this.props.locationGroups.map((value, index) => {    
+                    {this.props.logic.areas().map((value, index) => {    
                         return (
                         <LocationGroup
                             key={index}
                             groupName={value}
-                            locations={this.props.locations[value]}
+                            locations={this.props.logic.locationsForArea(value)}
                             expanded={this.props.expandedGroup === value}
                             handler={this.props.handleGroupClick}
                             locationHandler={this.props.handleLocationClick}
